+++ discO/core/__init__.py
@@ -11,13 +11,15 @@ __all__ = []
 # flatten structure
 
 # PROJECT-SPECIFIC
-from . import sample
+from . import fitter, sample
+from .fitter import *  # noqa: F401, F403
 from .measurement import *  # noqa: F403
 from .sample import *  # noqa: F403
 
 # alls
 __all__ += sample.__all__
 __all__ += measurement.__all__
+__all__ += fitter.__all__
 
 
 ##############################################################################